Refactored Common logic to single file for managing Category updates in parent form
